게시물 조회수 구현

+++ src/main/webapp/WEB-INF/jsp/usr/article/detail.jsp
@@ -1,5 +1,4 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-	pageEncoding="UTF-8"%>
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
 <c:set var="pageTitle" value="게시물 내용" />
 <%@ include file="../common/head.jspf"%>
@@ -29,6 +28,10 @@
 						<td>${article.extra__writerName}</td>
 					</tr>
 					<tr>
+						<th>조회수</th>
+						<td>${article.hitCount}</td>
+					</tr>
+					<tr>
 						<th>제목</th>
 						<td>${article.title}</td>
 					</tr>
@@ -41,12 +44,9 @@
 		</div>
 		<div class="btns">
 			<button class="btn-text-link" type="button" onclick="history.back();">뒤로가기</button>
-			<a class="btn-text-link" href="../article/modify?id=${article.id}">게시물
-				수정</a>
+			<a class="btn-text-link" href="../article/modify?id=${article.id}">게시물 수정</a>
 			<c:if test="${article.extra__actorCanDelete}">
-				<a class="btn-text-link"
-					onclick="if ( confirm('정말 삭제하시겠습니까?') == false) return false;"
-					href="../article/doDelete?id=${article.id}">게시물 삭제</a>
+				<a class="btn-text-link" onclick="if ( confirm('정말 삭제하시겠습니까?') == false) return false;" href="../article/doDelete?id=${article.id}">게시물 삭제</a>
 			</c:if>
 		</div>
 	</div>
